Revert "LPS-76246 Interface is not necessary. Convert it to a component without"
This reverts commit 78ff73589366c0feae7a280243cc12569b3cbf21.
